* 测试目录更改为D盘 * 加入对日志文件读取 * 修正全局日志存储路径错误 * 修正数据库原始操作类中的id获取范围 * 加入采集器多语言支持 * 日志模块加入独立文件写入设定和前置写入 * 修正用户模块一处逻辑错误 * 加入新的语言区块

+++ controller/database.go
@@ -87,7 +87,7 @@ func (this *Database) Update(table string, setField string, setValue string, id 
 }
 
 //Removes the specified id
-func (this *Database) Delete(table string, id int) (int64, error) {
+func (this *Database) Delete(table string, id int64) (int64, error) {
 	query := "delete from `" + table + "` where `id` = ?"
 	smat, err := this.db.Exec(query, id)
 	if err != nil {
